[#17] move from double to BigDecimal

+++ src/main/java/com/kubawach/nfs/core/model/Product.java
@@ -1,5 +1,7 @@
 package com.kubawach.nfs.core.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -11,5 +13,5 @@ public class Product {
         return new Product(source.getConcentration());
     }
 
-    private double concentration;
+    private BigDecimal concentration;
 }
